taller_mecanico/requirements.txt
django--3.2.4